Creating a module for lattice-related operations.
